requirements.txt
argcomplete==1.10.0
beautifulsoup4==4.8.0
certifi==2020.6.20
chardet==3.0.4
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.15
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
IMAPClient==2.1.0
joblib==0.16.0
langdetect==1.0.8
lxml==4.5.2
numpy==1.19.2
olefile==0.46
pandas==1.1.2
pdf2image==1.14.0
pdfminer.six==20181108
Pillow==7.2.0
py4j==0.10.7
pycryptodome==3.9.8
PyPDF2==1.26.0
pyspark==2.4.5
pytesseract==0.3.6
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
rfc3986==1.4.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.12.0
sklearn==0.0
sniffio==1.1.0
sortedcontainers==2.2.2
soupsieve==2.0.1
spark-nlp==2.6.1
SpeechRecognition==3.8.1
textract==1.6.3
threadpoolctl==2.1.0
tzlocal==1.5.1
xlrd==1.2.0
XlsxWriter==1.3.4
